整合vue稳定版-证书区分环境发送

+++ src/main/java/cn/com/cybertech/controller/AppInfoController.java
@@ -2,24 +2,20 @@ package cn.com.cybertech.controller;
 
 import cn.com.cybertech.model.AppInfo;
 import cn.com.cybertech.service.AppInfoService;
+import cn.com.cybertech.service.WebUserService;
 import cn.com.cybertech.tools.EncryptUtils;
 import cn.com.cybertech.tools.MessageCode;
 import cn.com.cybertech.tools.MessageCodeUtil;
 import cn.com.cybertech.tools.RestResponse;
 import cn.com.cybertech.tools.exception.ValueRuntimeException;
-import com.github.pagehelper.PageHelper;
-import com.github.pagehelper.PageInfo;
-import com.google.common.collect.Maps;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/app/info")
@@ -27,6 +23,9 @@ public class AppInfoController {
 
     @Autowired
     private AppInfoService appInfoService;
+
+    @Autowired
+    private WebUserService webUserService;
 
     @Autowired
     private MessageCodeUtil messageCodeUtil;
@@ -69,10 +68,14 @@ public class AppInfoController {
     //删除应用
     @RequestMapping("/deleteAppInfo")
     public RestResponse deleteAppInfo(Integer appId) {
+        String users = webUserService.getUserAppByAppId(appId);
+        if (StringUtils.isNotBlank(users)) {
+            return RestResponse.res(MessageCode.APPINFO_ERR_DEL, users+"绑定此应用，请解除用户绑定后删除");
+        }
         int msgCode = MessageCode.BASE_SUCC_CODE;
-        try{
+        try {
             appInfoService.deleteAppInfo(appId);
-        }catch (ValueRuntimeException e){
+        } catch (ValueRuntimeException e) {
             msgCode = (int) e.getValue();
         }
         return RestResponse.res(msgCode, messageCodeUtil.getMessage(msgCode));
